Rough idea on how to assure data integrity when working with bluetooth

+++ app/src/main/java/com/example/mdpandroid/util/RawImageParser.java
@@ -4,17 +4,19 @@ import android.content.Context;
 import android.util.Base64;
 import android.util.Log;
 
-import org.json.JSONArray;
+import androidx.annotation.Nullable;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.BufferedOutputStream;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashMap;
 
 /**
  * Created by Kenneth on 18/2/2020.
@@ -22,13 +24,82 @@ import java.io.OutputStreamWriter;
  */
 public class RawImageParser {
 
-    private JSONObject payload;
     private String imgB64;
+
+    private static boolean DEBUG = true;
 
     private static final String TAG = "ImgParse";
     private Context context;
 
+    private int imageCount = 1;
+    private int imagePacketCount = -1;
+    private HashMap<Integer, String> base64MapData;
+    private ArrayList<File> imageFiles;
+    private CallbackHandler mHandler;
+    private static final String MSG_ACK = "ACK";
+
+    public RawImageParser(Context context, CallbackHandler handler) {
+        this.context = context;
+        this.mHandler = handler;
+        this.imageFiles = new ArrayList<>();
+        this.base64MapData = new HashMap<>();
+    }
+
+    public void parseString(String s) {
+        try {
+            JSONObject payload = new JSONObject(s);
+            saveDbgFile(s, "payload-" + (new Date().toString()));
+
+            Log.d(TAG, "Processing " + s);
+            if (!(payload.has("state") || payload.has("c"))) {
+                Log.e(TAG, "Payload no state or packet number, discarding");
+                return; // Nothing to do
+            }
+
+            switch (payload.getString("state").toUpperCase()) {
+                case "S":
+                    Log.d(TAG, "Detected Start String, saving image packet count");
+                    this.imagePacketCount = payload.getInt("data");
+                    this.base64MapData.clear(); // We clear in case we are receiving another image
+                    this.mHandler.sendCommand(MSG_ACK);
+                    break;
+                case "E":
+                    Log.d(TAG, "Detected Ennd String, checking integrity");
+                    if (base64MapData.size() != imagePacketCount) resetAndResend();
+                    else processString();
+                    break;
+                case "D":
+                    this.base64MapData.put(payload.getInt("c"), payload.getString("data"));
+                    this.mHandler.sendCommand(MSG_ACK);
+                    break;
+            }
+
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void processString() {
+        Log.d(TAG, "Data integrity confirmed. Concatanating string");
+        this.imgB64 = "";
+        for (int i = 1; i <= this.imagePacketCount; i++) {
+            this.imgB64 += this.base64MapData.get(i);
+        }
+        this.mHandler.sendCommand(MSG_ACK);
+        saveImage();
+    }
+
+    private void resetAndResend() {
+        Log.d(TAG, "Data Integrity failed. discarding results and re-request for new image data");
+        this.imagePacketCount = -1;
+        this.imageCount = 1;
+        this.base64MapData.clear();
+        this.imageFiles.clear();
+        this.mHandler.sendCommand(Cmd.REQIMG);
+    }
+
     private void saveDbgFile(String s, String filename) {
+        if (!DEBUG) return;
         try {
             File f = new File(context.getFilesDir().getAbsoluteFile(), filename + ".txt");
             if (f.exists()) f.delete();
@@ -41,50 +112,40 @@ public class RawImageParser {
         }
     }
 
-    public RawImageParser(String payload, Context context) {
-        JSONObject tmpPayload = null;
-        this.context = context;
-        Log.d(TAG, "JSON String Length: " + payload.length());
-        saveDbgFile(payload, "dbg");
-
-        try {
-            tmpPayload = new JSONObject(payload);
-            this.payload = tmpPayload;
-
-            saveImage();
-        } catch (JSONException json) {
-            Log.d(TAG, "JSON Exception (" + json.getLocalizedMessage() + ")");
-            this.payload = null;
-        }
-    }
-
     private void saveImage() {
-        if (this.payload == null || context == null) return;
+        if (context == null) return;
 
-        Log.d(TAG, "Saving images");
-
+        Log.d(TAG, "Saving image #" + this.imageCount + " to file");
         try {
-            JSONArray arr = this.payload.getJSONArray("imgRaw");
-            for (int i = 0; i < arr.length(); i++) {
-                Log.d(TAG, "Processing image #" + i);
-                String img = arr.getString(i);
-                saveDbgFile(img, "predecocde_" + i);
-                byte[] imgBytes = Base64.decode(img, Base64.DEFAULT);
-                saveDbgFile(imgBytes.toString(), "postDecode" + i);
-                File f = new File(context.getFilesDir().getAbsoluteFile(), "img_" + i + ".jpg");
-                f.getParentFile().mkdirs();
-                BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(f));
-                bos.write(imgBytes);
-                bos.flush();
-                bos.close();
-                Log.d(TAG, "Image saved to " + f.getAbsolutePath());
-            }
-        } catch (JSONException ex) {
-            Log.d(TAG, "JSON Exception saveImage(): (" + ex.getLocalizedMessage() + ")");
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
+            saveDbgFile(imgB64, "predecode_" + this.imageCount);
+            byte[] imgBytes = Base64.decode(this.imgB64, Base64.DEFAULT);
+            saveDbgFile(imgBytes.toString(), "postDecode_" + this.imageCount);
+
+            File f = new File(context.getFilesDir().getAbsoluteFile(), "img_" + this.imageCount + ".jpg");
+            f.getParentFile().mkdirs();
+            if (f.exists()) f.delete();
+            BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(f));
+            bos.write(imgBytes);
+            bos.flush();
+            bos.close();
+            Log.d(TAG, "Image saved to " + f.getAbsolutePath());
+            this.imageCount++;
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+
+    public void stitchImages() {
+        // TODO: Get all images and stitch them together
+    }
+
+    @Nullable
+    public File getImageFile(int image) {
+        if (this.imageFiles.size() <= image) return null;
+        return this.imageFiles.get(image);
+    }
+
+    public interface CallbackHandler {
+        void sendCommand(String msg);
     }
 }
